\name{fracdiffR-package}
\alias{fracdiffR-package}
\alias{fracdiffR}
\docType{package}
\title{
\packageTitle{fracdiffR}
}
\description{
\packageDescription{fracdiffR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fracdiffR}
\packageIndices{fracdiffR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fracdiffR}

Maintainer: \packageMaintainer{fracdiffR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
